requirements.txt
Django==4.2.16
python-decouple==3.8
Pillow==10.4.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-extensions==3.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
openai==1.35.3
django-widget-tweaks==1.5.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.25
python-dotenv==1.0.0
requests==2.32.3
gunicorn==22.0.0
dj-database-url==2.1.0
psycopg[binary]
twilio==9.7.0
whitenoise==6.7.0
